MNT: fallback to v0 databroker
FXI needs to support swapping the handlers out very late to support
their way of getting the frame timestamps out (which is a feature we
intentionally did not move to v1/v2 because it is too fiddly).
Eventually we will need to do a database migration where we re-write
their events to put the timestamps as a data field, but not this week!